+++ internal/integration/cache_test_helpers.go
@@ -33,6 +33,7 @@ type CacheTestConfigVars struct {
 
 	// Generic variables.
 	Var1 string
+	Var2 string
 }
 
 // CachePreTestFn is an optional closure to be called before tests are run, this
@@ -75,6 +76,7 @@ func (e cacheTestEnvironment) RenderConfig() string {
 		"$ID", e.configVars.ID,
 		"$PORT", e.configVars.Port,
 		"$VAR1", e.configVars.Var1,
+		"$VAR2", e.configVars.Var2,
 	).Replace(e.configTemplate)
 }
 
@@ -122,6 +124,14 @@ func CacheTestOptVarOne(v string) CacheTestOptFunc {
 	}
 }
 
+// CacheTestOptVarTwo sets an arbitrary variable for the test that can be
+// injected into templated configs.
+func CacheTestOptVarTwo(v string) CacheTestOptFunc {
+	return func(env *cacheTestEnvironment) {
+		env.configVars.Var2 = v
+	}
+}
+
 // CacheTestOptPreTest adds a closure to be executed before each test.
 func CacheTestOptPreTest(fn CachePreTestFn) CacheTestOptFunc {
 	return func(env *cacheTestEnvironment) {
@@ -161,9 +171,6 @@ func (i CacheTestList) Run(t *testing.T, configTemplate string, opts ...CacheTes
 		env.ctx, done = context.WithTimeout(env.ctx, env.timeout)
 		t.Cleanup(done)
 
-		if env.preTest != nil {
-			env.preTest(t, env.ctx, env.configVars.ID, &env.configVars)
-		}
 		test.fn(t, &env)
 	}
 }
@@ -174,6 +181,10 @@ func namedCacheTest(name string, test cacheTestDefinitionFn) CacheTestDefinition
 	return CacheTestDefinition{
 		fn: func(t *testing.T, env *cacheTestEnvironment) {
 			t.Run(name, func(t *testing.T) {
+				t.Parallel()
+				if env.preTest != nil {
+					env.preTest(t, env.ctx, env.configVars.ID, &env.configVars)
+				}
 				test(t, env)
 			})
 		},